requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.10
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.3.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dill==0.4.0
dj-database-url==2.1.0
Django==5.0.8
django-cors-headers==4.3.1
django-extensions==4.1
django-filter==23.5
django-model-utils==5.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
graphviz==0.21
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
mccabe==0.7.0
packaging==25.0
Pillow==10.1.0
plantuml==0.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pydotplus==2.0.2
PyJWT==2.10.1
pylint==3.3.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
sqlparse==0.5.3
stripe==7.7.0
tomlkit==0.13.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
